ACL 70% complete fix some code bugs fix overall layout of project

+++ application/controllers/users_inform_controller.php
@@ -8,26 +8,31 @@
 		
 			function user_detail()
 			{
+				$id=$this->session->userdata('user_id');
+				$user_id=$id->id;
 				$this->load->model('users_inform_model');
-				$data=$this->users_inform_model->user_detail();
+				$data=$this->users_inform_model->user_detail($user_id);
 				$this->load->view('users_inform_view',['data'=>$data]);
 			}
 
 										// Add_USER_DETAIL
 			function add_user_inform(){
-				
+				$id=$this->session->userdata('user_id');
+				$user_id=$id->id;
 				$data=$this->input->post('');
 				$this->load->model('users_inform_model');
-				$this->users_inform_model->add_user_inform();
+				$this->users_inform_model->add_user_inform($user_id);
 
 			}
 
 			public function delete()
 			{
+				$id=$this->session->userdata('user_id');
+				$user_id=$id->id;
 				 $id= $this->input->post('id');
        
 
-      $this->db->where('id',$id)->delete('user_inform');
+      $this->db->where('user_id',$user_id)->where('id',$id)->delete('user_inform');
 			}
 
 			 function __construct() {
